DRY, 23% coverage
