+++ run_tests.py
@@ -11,6 +11,7 @@ sys.path.insert(0, source_dir)
 
 from tests.test_notification_pusher import NotificationPusherTestCase
 from tests.test_redirect_checker import RedirectCheckerTestCase
+from tests.test_utils import UtilsTestCase
 
 @contextmanager
 def mocked_connection():
@@ -26,6 +27,7 @@ if __name__ == '__main__':
     suite = unittest.TestSuite((
         unittest.makeSuite(NotificationPusherTestCase),
         unittest.makeSuite(RedirectCheckerTestCase),
+        unittest.makeSuite(UtilsTestCase),
     ))
     with mocked_connection():
         result = unittest.TextTestRunner().run(suite)
